move prompt dialog out of alertdialog + rename from confirm
